Proposal 7
Add a new rule with the wording "The voting for any specific proposal is to take place during a 1 hour period, henceforth referred to as the voting period. During this time votes for and against the proposal may be cast and retracted at will, but at the end of this voting period the current totals of votes cast for and against the proposal are what are counted for the majority vote meant by rule 1. Additionally, if at any point a proposal has a 2/3 majority as counted by the votes cast, it passes immediately."